Moved GradientTransform into its own source files.
