+++ Gems/GradientSignal/Code/Source/GradientTransform.cpp
